requirements.txt
feedparser>=6.0
openai>=1.0
jinja2>=3.1
pyyaml>=6.0
requests>=2.31
beautifulsoup4>=4.12
arxiv>=2.1
python-dotenv>=1.0
